minor optimization

+++ Figures/Pawn.h
@@ -13,7 +13,7 @@ class Pawn : public Figure {
 public:
     Pawn();
 
-    Pawn(bool side, Coordinates coordinates, uint8_t num);
+    Pawn(bool side, Coordinates coordinates, int num);
 
     void calculate_possible_moves(std::vector<Figure> white_figures, std::vector<Figure> black_figures) override;
 };
